% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{alpha_visualization}
\alias{alpha_visualization}
\title{The visualization of alpha diversity}
\usage{
alpha_visualization(x, y, index
}
\arguments{
\item{x}{x is phenotype cantains the subjects' characteristic;}

\item{y}{y is the result of diversity by diversity function;}

\item{index}{the type of diversity measures}
}
\value{
Return boxplot
}
\description{
alpha_visualization
The aim of alpha_visualization function is to display the alpha diversity by boxplot
}
\details{
10/01/2020  ShenZhen China
}
\examples{
 result <- alpha_visualization(phen, diversity_result$alpha, "diversity_shannon")

}
\author{
Hua Zou
}
