% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{CvLasso}
\alias{CvLasso}
\title{CvLasso
select feature & predict performance using lasso method}
\usage{
CvLasso(
  responsetag,
  response,
  microbio,
  transformM = "IQN",
  normalization = NULL,
  family,
  numfold,
  optimal,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
CvLasso
select feature & predict performance using lasso method
}
