% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_into_phyloseq.R
\name{data2phyloseq}
\alias{data2phyloseq}
\title{data2phyloseq}
\usage{
data2phyloseq(x, y, z, sampleid)
}
\arguments{
\item{x}{x  (Required) A table including multiple vectors of subjects' characteristic;}

\item{y}{y  (Required) A table including multiple metabolic measures of subjects;}

\item{z}{z  (Required) A table including multiple taxonomy, colnames are sampleid, rownames are taxonomy;}

\item{sampleid}{A  (Required) character of the subjects' ID to connect the three groups.}
}
\value{
A phyloseq object
}
\description{
Convert all the variate such as phenotype metadata and the relative abundance profiles in the data.frame/matrix into the data of phyloseq type
}
\details{
13/01/2020  ShenZhen China
}
\examples{
vignettes/example-data2phyloseq.R

}
\author{
Hua Zou
}
