% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{microbial diversity show the ecosystem state}
\usage{
diversity(x, y, tag, group)
}
\arguments{
\item{x}{x is phenotype cantains the subjects' characteristic;}

\item{y}{y is the overall profile, colnames are samplenames, rownames are mearsures;}

\item{tag}{which connect x and y.}

\item{group}{beta diversity in group}
}
\value{
Return diversity result
}
\description{
diversity
The aim of diversity function is to calculate the global indicators of the ecosystem state
}
\details{
10/01/2020  ShenZhen China
}
\examples{
 result <- diversity(phen, spf, "SampleID", "Group")

}
\author{
Hua Zou
}
