% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{filterEx}
\alias{filterEx}
\title{filterEx}
\usage{
filterEx(x, percent.outlier = 0.05)
}
\arguments{
\item{x}{vector}

\item{percent.outlier}{numberic,defalut 0.05}
}
\value{
vector
}
\description{
filter the extreme value
}
