% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{filterPer}
\alias{filterPer}
\title{filterPer}
\usage{
filterPer(x, row, percent, include = T)
}
\arguments{
\item{x}{, a data.frame object, species abundance profile}

\item{row, }{a numberic, 1/2}

\item{percent, }{percent to retain or remove}

\item{include, }{a logistic , T/F}
}
\value{
data.frame
}
\description{
filterPer
}
\examples{

}
