% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{A brief introduction to this function.}
\usage{
foo(x1 = 2, x2 = 21)
}
\arguments{
\item{x1}{explain what is x1}

\item{x2}{}
}
\value{
explain what is returned by this function.
}
\description{
A brief introduction to this function.
}
\examples{
foo()
}
