% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_estimating_equations.R
\name{gee_test}
\alias{gee_test}
\title{Generalized Estimating Equations}
\usage{
gee_test(physeq, variable, adjust_covariates, PID, pcutoff)
}
\arguments{
\item{physeq}{(Required).  A \code{phyloseq} object containing merged information of abundance,
sample data including the measured variables and categorical information of the samples.}

\item{variable}{(Required). A character string as the measurement (repeated responses).}

\item{adjust_covariates}{(optional). A set of Characters string that are covariates.
information.}

\item{PID}{(Required) the repeated subjects}
}
\value{
Returns a data.frame with the results of gee.
}
\description{
To estimate the relationship bwtween features/taxa and phenotypes.
}
\details{
15/01/2020  ShenZhen China
}
\examples{

library(dplyr)
library(geepack)
data("physeq_data")
physeq <- physeq_data
variable <- c("BMI", "Glycine")
adjust_covariates <- c("Age", "Sex")
PID <- "ID"
pcutoff <- 0.05

gee_test(physeq, variable, adjust_covariates, PID, pcutoff)

}
\author{
Hua Zou
}
