% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{LassoFeatruePlot}
\alias{LassoFeatruePlot}
\title{LassoFeaturePlot
plot the no-zeor lasso score of feature}
\usage{
LassoFeatruePlot(lassoresult, top = NULL, type, title, optimal)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
LassoFeaturePlot
plot the no-zeor lasso score of feature
}
