% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel_test.R
\name{mantel_test}
\alias{mantel_test}
\title{mantel_test}
\usage{
mantel_test(microbiota, metadata, methodf, methods)
}
\arguments{
\item{microbiota}{a data.frame object, species profile, row is sample}

\item{metadata}{a data.frame object, metadata inf}

\item{methodf}{feature space distance method}

\item{methods}{sample space distance method, vegdist}
}
\value{
figure
}
\description{
mantel_test
}
