% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{multiplot}
\alias{multiplot}
\title{multiplot
multi plot to combine}
\usage{
multiplot(plotlist, file, cols = 1, layout = NULL)
}
\arguments{
\item{layout}{}
}
\value{

}
\description{
multiplot
multi plot to combine
}
