% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_tran}
\alias{network_tran}
\title{Network_tran
Transform the matrix to ggraph structure}
\usage{
network_tran(adj, abun, rankinf)
}
\arguments{
\item{rankinf}{}
}
\value{

}
\description{
Network_tran
Transform the matrix to ggraph structure
}
