% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{ordination}
\alias{ordination}
\title{Ordination and beta dispersion}
\usage{
ordination(physeq, method, grouping_column)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{which_distance}{(Optional). A string character specifying dissimilarity index to be used in calculating pairwise distances (Default index is "bray".).
"unifrac","wunifrac","manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard", "gower", "altGower",
"morisita", "horn", "mountford", "raup" , "binomial", "chao", "cao" or "mahalanobis".}

\item{method}{(Optional). A character string specifying ordination method. All methods available to the \code{ordinate} function
of \code{phyloseq} are acceptable here as well ( Default is "PCOA").
"DCA", "CCA", "RDA", "CAP", "DPCoA", "NMDS", "MDS", "PCoA", "PCA", "Tsne".}

\item{grouping_column}{(Required). Character string specifying name of a categorical variable that is preffered for grouping the information.
information.}
}
\value{
Returns a list of three items: \itemize{
        \item A solution of the ordination
        \item A \code{data.frame} of betadispersion results; compared groups, corresponding pairwise observed p-values
        and significant labels.
        \item an object of class "adonis" with all components ; see \code{adonis} for details.
        }
}
\description{
This function performs ordination using a selected method. It also performs PERMANOVA using a distance matrix corresponding
to the taxa abundance under the different conditions/groups. It also computes  pairwise beta dispersion for
all conditions of a selected grouping variable, beta dispersion is measured as the average distance of group members to the group centroid.
The function a solution of ordination, PERMANOVA results and beta dispersion results. See value for details.
}
\details{
14/01/2020  ShenZhen China
}
\examples{
data(physeq_data)
physeq <- physeq_data
ord.res <- ordination(physeq, method = "NMDS", grouping_column = "Stage")


}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Hua Zou
}
