% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_foldchange.R
\name{pair_foldchange}
\alias{pair_foldchange}
\title{pair_foldchange
The fold change value between the pair samples}
\usage{
pair_foldchange(microbiota, metadata, percent, order)
}
\arguments{
\item{microbiota}{a dataframe object,species abundance, row is sample Id}

\item{metadata}{a dataframe object, metadata, row is sample Id}

\item{percent}{numberic, to cutoff value of species occranece}

\item{order}{logistic , to detemain the time order}
}
\value{
list include ggplot object and the data.frame of foldchange
}
\description{
pair_foldchange
The fold change value between the pair samples
}
