% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{pairNet}
\alias{pairNet}
\title{pairNet
Pair Networt for a specific species on 2 time point}
\usage{
pairNet(microbiota, metadata, cutoff, species, name, ...)
}
\arguments{
\item{microbiota}{a data.frame object, species profile, row is sample}

\item{metadata}{a data.frame object, metadata table, row is sample}

\item{cutoff}{numberic, the sparcc coefficient of species-species indicate
the cutoff of eage}

\item{species}{charater, species names, eg "Vibrio_harveyi"}

\item{name}{charater, provida a name for the specific species of treatment}
}
\value{
matrix
}
\description{
pairNet
Pair Networt for a specific species on 2 time point
}
