% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{perfPlot}
\alias{perfPlot}
\title{perfPlot
performance of randomForesttwo}
\usage{
perfPlot(randomRes, title)
}
\arguments{
\item{randomRes}{obeject of randomForesttwo}

\item{title}{charcter, title of figure}
}
\value{

}
\description{
perfPlot
performance of randomForesttwo
}
