% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{PERMANOVA1}
\alias{PERMANOVA1}
\title{Permutational Multivariate Analysis of Variance Using Distance Matrices}
\usage{
PERMANOVA1(physeq, sampleid)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
sample data including the measured variables and categorical information of the samples.}

\item{sampleid}{(Required) A character of the sampleid to connect phenotype and  profiles.}
}
\value{
permanova result
}
\description{
PERMANOVA1 The PERMANOVA1 aimed to assess the association between the overall profile and phenotype
}
\details{
09/01/2020  ShenZhen China
}
\examples{

data(physeq_data)
sampleid <- "SampleID"
PERMANOVA1(physeq_data, sampleid)

}
\author{
Hua Zou
}
