% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_central_distance.R
\name{plot_centrel}
\alias{plot_centrel}
\title{plot_centrel}
\usage{
plot_centrel(pro, method, config, color = time_colour)
}
\arguments{
\item{pro}{a data.frame or matrix, species abundance}

\item{method}{a character, distance method,eg "bray", "euclidean"}

\item{config}{a data.frame include sample metadata}

\item{color}{vector, group color}
}
\value{
ggplot object
}
\description{
plot_centrel
}
