% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ordination.R
\name{plot_ordination}
\alias{plot_ordination}
\title{Ordination plots}
\usage{
plot_ordination(physeq, method, grouping_column)
}
\arguments{
\item{ordination.res}{(Required).  A solution of ordination results returned from \link[microbiotaPair]{ordination}.}

\item{method}{(Optional). A character string specifying ordination method. All methods available to the \code{ordinate} function
of \code{phyloseq} are acceptable here as well ( Default is "PCOA").
"DCA", "CCA", "RDA", "CAP", "DPCoA", "NMDS", "MDS", "PCoA", "PCA", "Tsne".}

\item{grouping_column}{(Required). Character string specifying name of a categorical variable that is preffered for grouping the information.
information.}

\item{PID}{id (Required) for paired test}
}
\value{
Returns a ggplot object.
}
\description{
This function produces visualisation of ordination and beta dispersion results.
}
\details{
14/01/2020  ShenZhen China
}
\examples{

data(physeq_data)
physeq <- physeq_data
ord.res <- ordination(physeq, method="NMDS", grouping_column="Stage")
plot_ordination(ord.res, method="PCoA", grouping_column="Stage", PID="ID")


}
\references{
\url{http://userweb.eng.gla.ac.uk/umer.ijaz/}, Umer Ijaz, 2015
}
\author{
Hua Zou
}
