% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_distance.R
\name{plot_pair_distance}
\alias{plot_pair_distance}
\title{plot_pair_distance}
\usage{
plot_pair_distance(
  microbiota,
  metadata,
  method = "bray",
  percent.outlier = 0.05
)
}
\arguments{
\item{microbiota}{a dataframe object,species abundance, row is sample Id}

\item{metadata}{a dataframe object, metadata, row is sample Id}

\item{method}{a  character, distance method eg "bray"}

\item{percent.outlier}{a number, percent of the ourlier}
}
\value{
ggplot object
}
\description{
plot_pair_distance
}
