% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{randomForestTwo}
\alias{randomForestTwo}
\title{randomForestTwo
two class randomforest model}
\usage{
randomForestTwo(data, metadata, response, repeatNum, foldNum, factorLev)
}
\arguments{
\item{data}{data.frame or matrix,species data or metabolism data, row is sample ID}

\item{metadata}{data.frame or matrix, 0/1 response dataset, row is sample ID}

\item{response}{character, response feature name}

\item{repeatNum}{numberic, repeat number of feeature selection produce}

\item{foldNum}{numberic, cross validation}

\item{factorLev}{vector, 0/1 factor name}
}
\value{
list
}
\description{
randomForestTwo
two class randomforest model
}
