% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{randomForestTworobust}
\alias{randomForestTworobust}
\title{randomForestTworobust}
\usage{
randomForestTworobust(data, metadata, response, repeatNum, foldNum, factorLev)
}
\arguments{
\item{factorLev}{}
}
\value{

}
\description{
randomForestTworobust
}
