% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{repeatNetwork}
\alias{repeatNetwork}
\title{repeatNetwork
 to restruct the robust network using the resample method}
\usage{
repeatNetwork(dat, repeatN, sampleN)
}
\arguments{
\item{sampleN}{}
}
\value{

}
\description{
repeatNetwork
 to restruct the robust network using the resample method
}
