% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{simplenet}
\alias{simplenet}
\title{simpleNet}
\usage{
simplenet(adjmatrix, main, ...)
}
\arguments{
\item{adjmatrix}{the pairNet result}

\item{main}{the network title}
}
\value{
figure
}
\description{
simpleNet
}
\examples{

}
