% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{sparccNet}
\alias{sparccNet}
\title{sparcc network}
\usage{
sparccNet(dat = microbiota, cutoff = 0.3, main = "test", count = F)
}
\arguments{
\item{dat}{data.frame or matrix, row is sample ID}

\item{cutoff}{numberic, the sparcc coefficient of species-species indicate
the cutoff of eage}

\item{main}{character, the network title}

\item{count}{logistic, indicated the microbiota data is count of relative abundance}

\item{layout}{character, the paremeter from the igraph indicated the network layout}

\item{...}{}
}
\value{
figure
}
\description{
sparcc network
}
