% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_part_compare.R
\name{two_part_compare}
\alias{two_part_compare}
\title{two part compare
two part (paired test & McNemar’s Chi-Square Test) to maximum the data infomation}
\usage{
two_part_compare(microbiota, metadata, percent)
}
\arguments{
\item{microbiota}{a data.frame, microbiome data row is sample id, col is variable}

\item{metadata}{a data.frame, metadata row is sample id ,col is variable}

\item{percent}{numberic, 0-1, the cutoff of species occrance}
}
\value{
data.frame
}
\description{
two part compare
two part (paired test & McNemar’s Chi-Square Test) to maximum the data infomation
}
