% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilcoxon_Sign_Rank_Test.R
\name{wilcox_sign}
\alias{wilcox_sign}
\title{Wilcoxon Sign-Rank Test}
\usage{
wilcox_sign(x, y, DNAID, PID, GROUP, grp1=NULL, grp2=NULL)
}
\arguments{
\item{x}{x is data.frame with sampleID and group; sampleID connected to y}

\item{y}{y is table rownames->taxonomy; colnames->sampleID}

\item{DNAID}{names of sampleID to connect x and y}

\item{PID}{id for paired test}

\item{GROUP}{names of group information, only contain two levels if grp1 or grp2 haven't been provided}

\item{grp1}{one of groups to be converted into 0}

\item{grp2}{one of groups to be converted into 1}
}
\value{
Returns a result of Wilcoxon Sign-Rank Test

type:       kind of data

Block:      group information

Num:        number of group

P-value:    P by Wilcoxon Sign-Rank Test

FDR:        adjusted by BH

Enrichment: directory by median or directory by rank

Occurence:  occurence of two groups

median:     both or each group

rank:       each group

FDR:        adjusted P value by BH

Odds Ratio:     95% Confidence interval
}
\description{
Tests for the difference between two related variables; takes into account the magnitude and direction of difference
}
\details{
01/10/2020 ShenZhen China
}
\examples{
result <- wilcox_sign(phen, spf, "SampleID", "ID", "Stage", "Before", "After")
}
\author{
Hua Zou
}
