% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_into_phyloseq.R
\name{data2phyloseq}
\alias{data2phyloseq}
\title{data2phyloseq}
\usage{
data2phyloseq(x, y, z, sampleid)
}
\arguments{
\item{metadata}{x  (Required) A table including multiple vectors of subjects' characteristic;}

\item{micro}{y  (Required) A table including multiple taxonomy, colnames are sampleid, rownames are taxonomy;}

\item{occ}{A numberic of the occurrence above to restain}
}
\value{
A phyloseq object
}
\description{
Convert all the variate such as phenotype metadata and the relative abundance profiles in the data.frame/matrix into the data of phyloseq type
}
\details{
13/01/2020  ShenZhen China
}
\examples{
library(phyloseq)
library(dplyr)
library(data.table)
library(tibble)

metadata <- read.csv(system.file("extdata", "phenotype.csv", package = "microbiotaPair"))
micro <- read.table(system.file("extdata", "abundance.profile", 
    package = "microbiotaPair"), sep = "\t")
sampleid <- "SampleID"

physeq_data <- data2phyloseq(metadata, micro)
# print(physeq_data)

}
\author{
Hua Zou
}
