% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{PERMANOVA1}
\alias{PERMANOVA1}
\title{Permutational Multivariate Analysis of Variance Using Distance Matrices}
\usage{
PERMANOVA1(x, y, tag)
}
\arguments{
\item{x}{x is phenotype cantains the subjects' characteristic;}

\item{y}{y is the overall profile, colnames are samplenames, rownames are mearsures;}

\item{tag}{which connect x and y.}
}
\value{
Return the F test result
}
\description{
PERMANOVA1
The aim of PERMANOVA1 function is to asess the association between the overall profile and phenotype
}
\details{
09/01/2020  ShenZhen China
}
\examples{
 result <- PERMANOVA1(phen, spf, tag="SampleID")

}
\author{
Hua Zou
}
