% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{PERMANOVA1}
\alias{PERMANOVA1}
\title{A brief introduction to this function.}
\usage{
PERMANOVA1(x, y)
}
\arguments{
\item{x}{x is phenotype cantains the subjects' characteristic;}

\item{y}{y is the overall profile}
}
\value{
Return the F test result
}
\description{
PERMANOVA1
The aim of PERMANOVA1 function is to asess the association between the overall profile and phenotype
}
\details{
09/01/2020  ShenZhen China
}
\examples{
 result <- PERMANOVA1(phen, spf)

}
\author{
Hua Zou
}
