% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha_diversity.R
\name{plot_alpha_diversity}
\alias{plot_alpha_diversity}
\title{Plot alpha diversity}
\usage{
plot_alpha_diversity(alphares)
}
\arguments{
\item{alphares}{(Required). A result from \code{alpha_diversity} function
sample data including the measured variables and categorical information of the samples.}

\item{method}{(Required). A list of character strings specifying \code{method} to be used to calculate for alpha diversity
in the data. Available methods are: "t.test", "wilcox.test".}
}
\value{
Returns a ggplot object
}
\description{
This function calculates alpha diversity of provided community data using
selected method(s). It performs paired wilcox or t-test of diversity measures between groups
and outputs a plot for each of the selected methods(indices) annotated with significance levels.
}
\details{
22/01/2020  ShenZhen China
}
\examples{
library(dplyr)
library(ggplot2)
source("R/config.R")
data(physeq_data)
alphares <- alpha_diversity(physeq_data)
plot_alpha_diversity(physeq_data, method, grouping_column, pid)

}
\author{
Huahui Ren, zouhua
}
