% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha_diversity.R
\name{plot_alpha_diversity}
\alias{plot_alpha_diversity}
\title{Plot alpha diversity}
\usage{
plot_alpha_diversity(physeq, method, grouping_column, pid)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
sample data including the measured variables and categorical information of the samples.}

\item{method}{(Required). A list of character strings specifying \code{method} to be used to calculate for alpha diversity
in the data. Available methods are: "observed", "chao1", "simpson", "shannon", "evenness" and "all".}

\item{grouping_column}{(Required). A character string specifying the name of a categorical variable containing grouping information.}

\item{pid}{(Required) A character string for paired test}
}
\value{
Returns a ggplot object
}
\description{
This function calculates alpha diversity of provided community data using
selected method(s). It performs pair-wise ANOVA of diversity measures between groups
and outputs a plot for each of the selected methods(indices) annotated with significance levels.
}
\details{
14/01/2020  ShenZhen China
}
\examples{
library(dplyr)
library(ggplot2)
data(physeq_data)
method <- c("shannon", "simpson")
grouping_column <- "Stage"
pid <- "ID"
plot_alpha_diversity(physeq_data, method, grouping_column, pid)

}
\author{
Hua Zou
}
