% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilcoxon_Sign_Rank_Test.R
\name{wilcox_sign}
\alias{wilcox_sign}
\title{Wilcoxon Sign-Rank Test}
\usage{
wilcox_sign(physeq, sampleid, PID, GROUP, grp1, grp2)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing merged information of abundance,
sample data including the measured variables and categorical information of the samples.}

\item{PID}{id (Required) for paired test}

\item{GROUP}{names (Required) of group information, only contain two levels if grp1 or grp2 haven't been provided}

\item{grp1}{one of groups to be converted into 0 (optional)}

\item{grp2}{one of groups to be converted into 1 (optional)}

\item{sampleid}{(Required) A character of the sampleid to connect phenotype and  profiles.}
}
\value{
Returns a result of Wilcoxon Sign-Rank Test

type:       kind of data

Block:      group information

Num:        number of group

P-value:    P by Wilcoxon Sign-Rank Test

FDR:        adjusted by BH

Enrichment: directory by median or directory by rank

Occurence:  occurence of two groups

median:     both or each group

rank:       each group

FDR:        adjusted P value by BH

Odds Ratio:     95% Confidence interval
}
\description{
Tests for the difference between two related variables; takes into account the magnitude and direction of difference
}
\details{
01/10/2020 ShenZhen China
}
\examples{

data(physeq_data)
sampleid <- "SampleID"
pid <- "ID"
group <- "Stage"
grp1 <- "Before"
grp2 <- "After"

result <- wilcox_sign(physeq_data, sampleid, pid, group, grp1, grp2)


}
\author{
Hua Zou/ Huahui Ren
}
